refactor: move getWorkdayData function to route.ts and remove queries.js
